Mark CommitProxyInterface::getHealthMetrics as legacy
