+++ fdbclient/include/fdbclient/CommitProxyInterface.h
@@ -50,7 +50,9 @@ struct CommitProxyInterface {
 	RequestStream<ReplyPromise<Void>> waitFailure;
 
 	RequestStream<struct TxnStateRequest> txnState;
-	RequestStream<struct GetHealthMetricsRequest> getHealthMetrics;
+	// Reserved to preserve the historical adjusted-endpoint numbering for the
+	// commit proxy interface. Health metric requests are served by GrvProxyInterface.
+	RequestStream<struct GetHealthMetricsRequest> legacyGetHealthMetrics;
 	RequestStream<struct ProxySnapRequest> proxySnapReq;
 	RequestStream<struct ExclusionSafetyCheckRequest> exclusionSafetyCheckReq;
 	RequestStream<struct GetDDMetricsRequest> getDDMetrics;
@@ -76,7 +78,7 @@ struct CommitProxyInterface {
 			    RequestStream<struct GetStorageServerRejoinInfoRequest>(commit.getEndpoint().getAdjustedEndpoint(3));
 			waitFailure = RequestStream<ReplyPromise<Void>>(commit.getEndpoint().getAdjustedEndpoint(4));
 			txnState = RequestStream<struct TxnStateRequest>(commit.getEndpoint().getAdjustedEndpoint(5));
-			getHealthMetrics =
+			legacyGetHealthMetrics =
 			    RequestStream<struct GetHealthMetricsRequest>(commit.getEndpoint().getAdjustedEndpoint(6));
 			proxySnapReq = RequestStream<struct ProxySnapRequest>(commit.getEndpoint().getAdjustedEndpoint(7));
 			exclusionSafetyCheckReq =
@@ -98,7 +100,7 @@ struct CommitProxyInterface {
 		streams.push_back(getStorageServerRejoinInfo.getReceiver(TaskPriority::ProxyStorageRejoin));
 		streams.push_back(waitFailure.getReceiver());
 		streams.push_back(txnState.getReceiver());
-		streams.push_back(getHealthMetrics.getReceiver());
+		streams.push_back(legacyGetHealthMetrics.getReceiver());
 		streams.push_back(proxySnapReq.getReceiver());
 		streams.push_back(exclusionSafetyCheckReq.getReceiver());
 		streams.push_back(getDDMetrics.getReceiver());
@@ -360,46 +362,6 @@ struct TxnStateRequest {
 	template <class Ar>
 	void serialize(Ar& ar) {
 		serializer(ar, data, sequence, last, broadcastInfo, reply, arena);
-	}
-};
-
-struct GetHealthMetricsReply {
-	constexpr static FileIdentifier file_identifier = 11544290;
-	Standalone<StringRef> serialized;
-	HealthMetrics healthMetrics;
-
-	explicit GetHealthMetricsReply(const HealthMetrics& healthMetrics = HealthMetrics())
-	  : healthMetrics(healthMetrics) {
-		update(healthMetrics, true, true);
-	}
-
-	void update(const HealthMetrics& healthMetrics, bool detailedInput, bool detailedOutput) {
-		this->healthMetrics.update(healthMetrics, detailedInput, detailedOutput);
-		BinaryWriter bw(IncludeVersion());
-		bw << this->healthMetrics;
-		serialized = bw.toValue();
-	}
-
-	template <class Ar>
-	void serialize(Ar& ar) {
-		serializer(ar, serialized);
-		if (ar.isDeserializing) {
-			BinaryReader br(serialized, IncludeVersion());
-			br >> healthMetrics;
-		}
-	}
-};
-
-struct GetHealthMetricsRequest {
-	constexpr static FileIdentifier file_identifier = 11403900;
-	ReplyPromise<struct GetHealthMetricsReply> reply;
-	bool detailed;
-
-	explicit GetHealthMetricsRequest(bool detailed = false) : detailed(detailed) {}
-
-	template <class Ar>
-	void serialize(Ar& ar) {
-		serializer(ar, reply, detailed);
 	}
 };
