+++ fdbserver/workloads/PrivateEndpoints.cpp
@@ -98,7 +98,7 @@ struct PrivateEndpoints : TestWorkload {
 		// addTestFor(&CommitProxyInterface::getStorageServerRejoinInfo);
 		addTestFor(&CommitProxyInterface::waitFailure);
 		// addTestFor(&CommitProxyInterface::txnState);
-		// addTestFor(&CommitProxyInterface::getHealthMetrics);
+		// addTestFor(&CommitProxyInterface::legacyGetHealthMetrics);
 		// addTestFor(&CommitProxyInterface::proxySnapReq);
 		addTestFor(&CommitProxyInterface::exclusionSafetyCheckReq);
 		// addTestFor(&CommitProxyInterface::getDDMetrics);
